Add doubleInputDialog method to Dialogs class and use it instead of the default Input Dialog when inserting a link (to make it possible to specify link text)

+++ Utilities/VersionData.java
@@ -37,7 +37,7 @@ public class VersionData {
 
     public static final String SETTINGS_LOCATION = getConfigFileLocation();
 
-    public static final int CONFIG_VERSION = 1;
+    public static final double CONFIG_VERSION = 1;
 
     /**
      * Get the location of Notepad config file
